+++ infra/modules/gcp-psoxy-rest/main.tf
@@ -37,16 +37,8 @@ locals {
   }
 }
 
-data "google_client_openid_userinfo" "me" {
-
-}
-
-locals {
-  # hacky way to determine if Terraform running as a service account or not
-  tf_is_service_account = endswith(data.google_client_openid_userinfo.me.email, "iam.gserviceaccount.com")
-
-  tf_qualifier          = local.tf_is_service_account ? "serviceAccount:" : "user:"
-  tf_principal          = "${local.tf_qualifier}${data.google_client_openid_userinfo.me.email}"
+module "tf_runner" {
+  source = "../../modules/gcp-tf-runner"
 }
 
 data "google_service_account" "function" {
@@ -57,7 +49,7 @@ data "google_service_account" "function" {
 # to provision Cloud Function, TF must be able to act as the service account that the function will
 # run as
 resource "google_service_account_iam_member" "act_as" {
-  member             = local.tf_principal
+  member             = module.tf_runner.iam_principal
   role               = "roles/iam.serviceAccountUser"
   service_account_id = data.google_service_account.function.id
 }